Moved join search code into its own file
